refactor(profile): extract profile-setup form sections into widgets (#4512)
Second commit of the profile_setup_screen.dart decomposition (epic #4339).
Behavior-preserving, UI-only: pulls the inline form out of the View so every
file in screens/profile_setup/ is under 400 lines.

Extracted into widgets/:
- ProfileAvatarSection (avatar preview + image-source buttons), with the
  platform picker plumbing in profile_image_picker.dart and the manual-URL
  sheet; an _ImageSourceButton dedupes the three source buttons.
- DisplayNameField / BioField / WebsiteField / UsernameField — each owns its
  focus node and rebuilds locally (focus-coloured label, bio counter). The
  name field takes the View's focus node so Save can focus it when empty.
  Shared underline/hint decorations live in profile_setup_field_decorations.dart.
- ProfileSetupListeners — the MultiBlocListener (profile-load -> controllers,
  save status snackbars/dialogs/navigation, avatar/banner upload status, and
  the verifier launch) plus the relay-reconnect retry.
- verifier_flow.dart holds launchVerifierFlow + VerifierRoutePusher.

The View is now a thin composition (controllers + lifecycle + Scaffold of the
section widgets); profile_setup_view.dart drops from ~1,400 to 279 lines.
State ownership is unchanged: name/bio/website/username stay controller-owned
and are read by Save; picture/banner stay in ProfileEditorState.
