+++ mobile/lib/screens/profile_setup/view/profile_setup_view.dart
@@ -1,30 +1,14 @@
-import 'dart:async';
-import 'dart:io';
-
 import 'package:divine_ui/divine_ui.dart';
-import 'package:file_selector/file_selector.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
-import 'package:image_picker/image_picker.dart';
-import 'package:nostr_app_bridge_repository/nostr_app_bridge_repository.dart';
 import 'package:openvine/blocs/my_profile/my_profile_bloc.dart';
 import 'package:openvine/blocs/profile_editor/profile_editor_bloc.dart';
 import 'package:openvine/l10n/l10n.dart';
 import 'package:openvine/providers/app_providers.dart';
-import 'package:openvine/providers/nostr_client_provider.dart';
-import 'package:openvine/providers/user_profile_providers.dart';
-import 'package:openvine/screens/apps/web_iframe_sandbox_screen.dart';
 import 'package:openvine/screens/profile_setup/widgets/widgets.dart';
 import 'package:openvine/widgets/profile/nostr_info_sheet_content.dart';
-import 'package:openvine/widgets/profile_editor/username_status_indicator.dart';
-import 'package:openvine/widgets/user_avatar.dart';
-import 'package:unified_logger/unified_logger.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 class ProfileSetupScreenView extends ConsumerStatefulWidget {
   const ProfileSetupScreenView({required this.isNewUser, super.key});
@@ -40,15 +24,11 @@ class _ProfileSetupScreenViewState extends ConsumerState<ProfileSetupScreenView>
   final _nameController = TextEditingController();
   final _bioController = TextEditingController();
   final _websiteController = TextEditingController();
-  final _pictureController = TextEditingController();
   final _nip05Controller = TextEditingController();
-  final ImagePicker _picker = ImagePicker();
+  // Owned here (not in DisplayNameField) so the Save action can focus the
+  // name field when it is left empty.
+  final _nameFocusNode = FocusNode();
   bool _refreshProfileOnResume = false;
-
-  static const _kInputBorder = UnderlineInputBorder(
-    borderRadius: BorderRadius.zero,
-    borderSide: BorderSide(color: VineTheme.neutral10),
-  );
 
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) {
@@ -61,61 +41,20 @@ class _ProfileSetupScreenViewState extends ConsumerState<ProfileSetupScreenView>
     context.read<MyProfileBloc>().add(const MyProfileFetchRequested());
   }
 
-  static const _kHintStyle = TextStyle(color: VineTheme.lightText);
-
-  // Focus nodes for tracking field focus state
-  final _nameFocusNode = FocusNode();
-  final _bioFocusNode = FocusNode();
-  final _websiteFocusNode = FocusNode();
-  final _usernameFocusNode = FocusNode();
-
-  // Local-preview state for the brief window between picking and the
-  // bloc receiving the staged URL. The bloc's `pendingPictureUrl` takes
-  // over once `pendingAvatarStatus == staged`. Cleared on the next pick
-  // or on `ProfilePictureUploadCleared`.
-  //
-  // Native picks: file path flows through the platform-channel EXIF
-  // stripper inside the upload service. Web picks: `image_picker` returns
-  // a blob URL `dart:io File` cannot open, so the bytes drive both the
-  // preview and the bytes-based upload path.
-  File? _selectedImage;
-  Uint8List? _selectedImageBytes;
-
   @override
   void initState() {
     super.initState();
     WidgetsBinding.instance.addObserver(this);
-    // Rebuild when display name changes so save button updates.
-    _nameController.addListener(_onFocusChange);
-    _websiteFocusNode.addListener(_onFocusChange);
-    // Add focus listeners to update label colors
-    _nameFocusNode.addListener(_onFocusChange);
-    _bioFocusNode.addListener(_onFocusChange);
-    _usernameFocusNode.addListener(_onFocusChange);
-  }
-
-  void _onFocusChange() {
-    setState(() {});
   }
 
   @override
   void dispose() {
     WidgetsBinding.instance.removeObserver(this);
-    _nameController.removeListener(_onFocusChange);
     _nameController.dispose();
     _bioController.dispose();
     _websiteController.dispose();
-    _pictureController.dispose();
     _nip05Controller.dispose();
-    _nameFocusNode.removeListener(_onFocusChange);
-    _bioFocusNode.removeListener(_onFocusChange);
-    _websiteFocusNode.removeListener(_onFocusChange);
-    _usernameFocusNode.removeListener(_onFocusChange);
     _nameFocusNode.dispose();
-    _bioFocusNode.dispose();
-    _websiteFocusNode.dispose();
-    _usernameFocusNode.dispose();
-
     super.dispose();
   }
 
@@ -123,286 +62,13 @@ class _ProfileSetupScreenViewState extends ConsumerState<ProfileSetupScreenView>
   Widget build(BuildContext context) {
     final pubkey = ref.watch(authServiceProvider).currentPublicKeyHex;
 
-    return MultiBlocListener(
-      listeners: [
-        BlocListener<MyProfileBloc, MyProfileState>(
-          listenWhen: (prev, curr) => curr is MyProfileLoaded,
-          listener: (context, myProfileState) {
-            if (myProfileState is! MyProfileLoaded) return;
-
-            final profile = myProfileState.profile;
-            final extractedUsername = myProfileState.extractedUsername;
-            final externalNip05 = myProfileState.externalNip05;
-
-            setState(() {
-              _nameController.text = profile.displayName ?? profile.name ?? '';
-              _bioController.text = profile.about ?? '';
-              _websiteController.text = profile.website ?? '';
-              _pictureController.text = profile.picture ?? '';
-
-              if (extractedUsername != null) {
-                _nip05Controller.text = extractedUsername;
-              }
-            });
-
-            final editorBloc = context.read<ProfileEditorBloc>();
-            // Seed bloc with the persisted picture so the avatar widget can
-            // render `pendingPictureUrl ?? persistedPictureUrl` purely from
-            // state, no widget-local fallback for the existing avatar.
-            editorBloc
-              ..add(InitialPersistedPictureSet(profile.picture))
-              ..add(InitialPersistedBannerSet(profile.banner));
-            if (extractedUsername != null) {
-              editorBloc.add(InitialUsernameSet(extractedUsername));
-            }
-            if (externalNip05 != null) {
-              // External NIP-05 now lives on Settings -> Nostr -> NIP-05.
-              // Seed editor state here so Save from Edit Profile preserves it.
-              editorBloc
-                ..add(InitialExternalNip05Set(externalNip05))
-                ..add(const Nip05ModeChanged(Nip05Mode.external_))
-                ..add(ExternalNip05Changed(externalNip05));
-            }
-          },
-        ),
-        BlocListener<ProfileEditorBloc, ProfileEditorState>(
-          listenWhen: (prev, curr) => prev.status != curr.status,
-          listener: (context, state) {
-            if (state.status == ProfileEditorStatus.success) {
-              // Invalidate profile providers so profile screen refetches
-              final currentPubkey = ref
-                  .read(authServiceProvider)
-                  .currentPublicKeyHex;
-              if (currentPubkey != null) {
-                ref.invalidate(fetchUserProfileProvider(currentPubkey));
-                ref.invalidate(userProfileReactiveProvider(currentPubkey));
-              }
-
-              ScaffoldMessenger.of(context).showSnackBar(
-                SnackBar(
-                  content: Row(
-                    children: [
-                      Container(
-                        padding: const EdgeInsets.all(8),
-                        decoration: const BoxDecoration(
-                          color: VineTheme.vineGreen,
-                          shape: BoxShape.circle,
-                        ),
-                        child: const DivineIcon(
-                          icon: DivineIconName.check,
-                          color: VineTheme.whiteText,
-                          size: 17,
-                        ),
-                      ),
-                      const SizedBox(width: 12),
-                      Text(
-                        context.l10n.profileSetupProfilePublished,
-                        style: const TextStyle(color: VineTheme.vineGreen),
-                      ),
-                    ],
-                  ),
-                  backgroundColor: VineTheme.whiteText,
-                ),
-              );
-              if (widget.isNewUser) {
-                Navigator.of(context).popUntil((route) => route.isFirst);
-              } else {
-                if (context.canPop()) {
-                  context.pop(true);
-                } else {
-                  context.go('/');
-                }
-              }
-            } else if (state.status ==
-                ProfileEditorStatus.confirmationRequired) {
-              // Show confirmation dialog for blank profile overwrite
-              showDialog<void>(
-                context: context,
-                builder: (dialogContext) => AlertDialog(
-                  backgroundColor: VineTheme.cardBackground,
-                  title: Text(
-                    context.l10n.profileSetupCreateNewProfile,
-                    style: const TextStyle(color: VineTheme.whiteText),
-                  ),
-                  content: Text(
-                    context.l10n.profileSetupNoExistingProfile,
-                    style: const TextStyle(color: VineTheme.secondaryText),
-                  ),
-                  actions: [
-                    TextButton(
-                      onPressed: () => Navigator.of(dialogContext).pop(),
-                      child: Text(
-                        context.l10n.profileCancelButton,
-                        style: const TextStyle(color: VineTheme.lightText),
-                      ),
-                    ),
-                    TextButton(
-                      onPressed: () {
-                        Navigator.of(dialogContext).pop();
-                        context.read<ProfileEditorBloc>().add(
-                          const ProfileSaveConfirmed(),
-                        );
-                      },
-                      child: Text(
-                        context.l10n.profileSetupPublishButton,
-                        style: const TextStyle(color: VineTheme.vineGreen),
-                      ),
-                    ),
-                  ],
-                ),
-              );
-            } else if (state.status == ProfileEditorStatus.failure) {
-              // Invalidate profile providers after rollback
-              final currentPubkey = ref
-                  .read(authServiceProvider)
-                  .currentPublicKeyHex;
-              if (currentPubkey != null) {
-                ref.invalidate(fetchUserProfileProvider(currentPubkey));
-                ref.invalidate(userProfileReactiveProvider(currentPubkey));
-              }
-              switch (state.error) {
-                case ProfileEditorError.usernameTaken:
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    SnackBar(
-                      content: Text(context.l10n.profileSetupUsernameTaken),
-                      backgroundColor: VineTheme.error,
-                      duration: const Duration(seconds: 3),
-                    ),
-                  );
-                case ProfileEditorError.usernameReserved:
-                  final username = state.username;
-                  showDialog<void>(
-                    context: context,
-                    builder: (context) => UsernameReservedDialog(username),
-                  );
-                case ProfileEditorError.claimFailed:
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    SnackBar(
-                      content: Text(context.l10n.profileSetupClaimFailed),
-                      backgroundColor: VineTheme.error,
-                    ),
-                  );
-                case ProfileEditorError.publishFailed:
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    SnackBar(
-                      content: Text(context.l10n.profileSetupPublishFailed),
-                      backgroundColor: VineTheme.error,
-                    ),
-                  );
-                case ProfileEditorError.noRelaysConnected:
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    SnackBar(
-                      content: Text(context.l10n.profileSetupNoRelaysConnected),
-                      backgroundColor: VineTheme.error,
-                      duration: const Duration(seconds: 6),
-                      action: pubkey == null
-                          ? null
-                          : SnackBarAction(
-                              label: context.l10n.profileSetupRetryLabel,
-                              textColor: VineTheme.whiteText,
-                              onPressed: () =>
-                                  _retryAfterRelayReconnect(context, pubkey),
-                            ),
-                    ),
-                  );
-                case null:
-                  break;
-              }
-            }
-          },
-        ),
-        BlocListener<ProfileEditorBloc, ProfileEditorState>(
-          listenWhen: (prev, curr) =>
-              prev.pendingAvatarStatus != curr.pendingAvatarStatus,
-          listener: (context, state) {
-            switch (state.pendingAvatarStatus) {
-              case PendingAvatarStatus.staged:
-                // Avatar preview has already swapped (BlocBuilder is rebuilding
-                // from `effectivePictureUrl`). The snackbar makes the staged
-                // contract explicit: bytes uploaded, not yet published.
-                ScaffoldMessenger.of(context).showSnackBar(
-                  SnackBar(
-                    content: Text(context.l10n.profileSetupUploadStaged),
-                    backgroundColor: VineTheme.vineGreen,
-                  ),
-                );
-              case PendingAvatarStatus.failed:
-                final classified =
-                    state.avatarUploadError ?? AvatarUploadError.generic;
-                ScaffoldMessenger.of(context).showSnackBar(
-                  SnackBar(
-                    content: Text(
-                      profileSetupUploadErrorMessage(context.l10n, classified),
-                    ),
-                    backgroundColor: VineTheme.error,
-                  ),
-                );
-              case PendingAvatarStatus.idle:
-              case PendingAvatarStatus.uploading:
-                break;
-            }
-          },
-        ),
-        BlocListener<ProfileEditorBloc, ProfileEditorState>(
-          listenWhen: (prev, curr) =>
-              prev.pendingBannerStatus != curr.pendingBannerStatus,
-          listener: (context, state) {
-            switch (state.pendingBannerStatus) {
-              case PendingBannerStatus.staged:
-                ScaffoldMessenger.of(context).showSnackBar(
-                  SnackBar(
-                    content: Text(
-                      context.l10n.profileSetupBannerUploadSuccess,
-                    ),
-                    backgroundColor: VineTheme.vineGreen,
-                  ),
-                );
-              case PendingBannerStatus.failed:
-                final classified =
-                    state.bannerUploadError ?? BannerUploadError.generic;
-                ScaffoldMessenger.of(context).showSnackBar(
-                  SnackBar(
-                    content: Text(
-                      profileSetupBannerUploadErrorMessage(
-                        context.l10n,
-                        classified,
-                      ),
-                    ),
-                    backgroundColor: VineTheme.error,
-                  ),
-                );
-              case PendingBannerStatus.idle:
-              case PendingBannerStatus.uploading:
-                break;
-            }
-          },
-        ),
-        BlocListener<ProfileEditorBloc, ProfileEditorState>(
-          listenWhen: (prev, curr) =>
-              prev.verifierStatus != curr.verifierStatus &&
-              curr.verifierStatus == VerifierStatus.launchRequested,
-          listener: (context, state) async {
-            final launched = await launchVerifierFlow(
-              editorBloc: context.read<ProfileEditorBloc>(),
-              myProfileBloc: context.read<MyProfileBloc>(),
-              pushVerifierRoute: (location, {extra}) async {
-                await context.push(location, extra: extra);
-              },
-            );
-            if (launched && !kIsWeb && mounted) {
-              _refreshProfileOnResume = true;
-            }
-            if (!launched && context.mounted) {
-              ScaffoldMessenger.of(context).showSnackBar(
-                SnackBar(
-                  content: Text(context.l10n.relaySettingsCouldNotOpenBrowser),
-                  backgroundColor: VineTheme.error,
-                ),
-              );
-            }
-          },
-        ),
-      ],
+    return ProfileSetupListeners(
+      isNewUser: widget.isNewUser,
+      nameController: _nameController,
+      bioController: _bioController,
+      websiteController: _websiteController,
+      nip05Controller: _nip05Controller,
+      onNativeVerifierLaunched: () => _refreshProfileOnResume = true,
       child: BlocBuilder<ProfileEditorBloc, ProfileEditorState>(
         builder: (context, profileEditorState) {
           return Scaffold(
@@ -463,353 +129,21 @@ class _ProfileSetupScreenViewState extends ConsumerState<ProfileSetupScreenView>
                           child: Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              // Profile Picture Section with overlapping buttons
-                              Center(
-                                child: SizedBox(
-                                  // 144 avatar + 20 (half of 40px buttons extending below)
-                                  height: 164,
-                                  width: 144,
-                                  child: BlocBuilder<ProfileEditorBloc, ProfileEditorState>(
-                                    buildWhen: (prev, curr) =>
-                                        prev.pendingAvatarStatus !=
-                                            curr.pendingAvatarStatus ||
-                                        prev.pendingPictureUrl !=
-                                            curr.pendingPictureUrl ||
-                                        prev.persistedPictureUrl !=
-                                            curr.persistedPictureUrl,
-                                    builder: (context, editorState) {
-                                      final isUploadingImage =
-                                          editorState.pendingAvatarStatus ==
-                                          PendingAvatarStatus.uploading;
-                                      return Stack(
-                                        clipBehavior: Clip.none,
-                                        children: [
-                                          // Profile picture preview
-                                          UserAvatar(
-                                            imageProvider:
-                                                _buildProfilePictureProvider(
-                                                  editorState,
-                                                ),
-                                            name: _nameController.text.trim(),
-                                            placeholderSeed: pubkey,
-                                            size: 144,
-                                            semanticLabel: context
-                                                .l10n
-                                                .profileSetupProfilePicturePreview,
-                                          ),
-                                          // Upload progress indicator
-                                          if (isUploadingImage)
-                                            Positioned(
-                                              top: 0,
-                                              left: 0,
-                                              width: 144,
-                                              height: 144,
-                                              child: DecoratedBox(
-                                                decoration: BoxDecoration(
-                                                  borderRadius:
-                                                      BorderRadius.circular(56),
-                                                  color: VineTheme
-                                                      .backgroundColor
-                                                      .withValues(alpha: 0.7),
-                                                ),
-                                                child: const Center(
-                                                  child:
-                                                      CircularProgressIndicator(
-                                                        color:
-                                                            VineTheme.vineGreen,
-                                                        strokeWidth: 3,
-                                                      ),
-                                                ),
-                                              ),
-                                            ),
-                                          // Image source buttons - overlapping bottom of avatar
-                                          Positioned(
-                                            bottom: 0,
-                                            left: 0,
-                                            right: 0,
-                                            child: Row(
-                                              mainAxisAlignment:
-                                                  MainAxisAlignment.center,
-                                              children: [
-                                                // Show camera button on mobile only
-                                                if (!_isDesktopPlatform()) ...[
-                                                  GestureDetector(
-                                                    onTap: isUploadingImage
-                                                        ? null
-                                                        : () => _pickImage(
-                                                            ImageSource.camera,
-                                                          ),
-                                                    child: Container(
-                                                      width: 40,
-                                                      height: 40,
-                                                      decoration: BoxDecoration(
-                                                        color: VineTheme
-                                                            .surfaceContainer,
-                                                        borderRadius:
-                                                            BorderRadius.circular(
-                                                              16,
-                                                            ),
-                                                        border: Border.all(
-                                                          color: VineTheme
-                                                              .outlineMuted,
-                                                          width: 2,
-                                                        ),
-                                                      ),
-                                                      child: Center(
-                                                        child: SvgPicture.asset(
-                                                          DivineIconName
-                                                              .cameraPlus
-                                                              .assetPath,
-                                                          width: 24,
-                                                          height: 24,
-                                                          colorFilter:
-                                                              const ColorFilter.mode(
-                                                                VineTheme
-                                                                    .primary,
-                                                                BlendMode.srcIn,
-                                                              ),
-                                                        ),
-                                                      ),
-                                                    ),
-                                                  ),
-                                                  const SizedBox(width: 12),
-                                                ],
-                                                GestureDetector(
-                                                  onTap: isUploadingImage
-                                                      ? null
-                                                      : () => _pickImage(
-                                                          ImageSource.gallery,
-                                                        ),
-                                                  child: Container(
-                                                    width: 40,
-                                                    height: 40,
-                                                    decoration: BoxDecoration(
-                                                      color: VineTheme
-                                                          .surfaceContainer,
-                                                      borderRadius:
-                                                          BorderRadius.circular(
-                                                            16,
-                                                          ),
-                                                      border: Border.all(
-                                                        color: VineTheme
-                                                            .outlineMuted,
-                                                        width: 2,
-                                                      ),
-                                                    ),
-                                                    child: Center(
-                                                      child: SvgPicture.asset(
-                                                        DivineIconName
-                                                            .imagesSquare
-                                                            .assetPath,
-                                                        width: 24,
-                                                        height: 24,
-                                                        colorFilter:
-                                                            const ColorFilter.mode(
-                                                              VineTheme.primary,
-                                                              BlendMode.srcIn,
-                                                            ),
-                                                      ),
-                                                    ),
-                                                  ),
-                                                ),
-                                                const SizedBox(width: 12),
-                                                // URL input button
-                                                GestureDetector(
-                                                  onTap: isUploadingImage
-                                                      ? null
-                                                      : () =>
-                                                            _showImageUrlSheet(
-                                                              context,
-                                                            ),
-                                                  child: Container(
-                                                    width: 40,
-                                                    height: 40,
-                                                    decoration: BoxDecoration(
-                                                      color: VineTheme
-                                                          .surfaceContainer,
-                                                      borderRadius:
-                                                          BorderRadius.circular(
-                                                            16,
-                                                          ),
-                                                      border: Border.all(
-                                                        color: VineTheme
-                                                            .outlineMuted,
-                                                        width: 2,
-                                                      ),
-                                                    ),
-                                                    child: Center(
-                                                      child: SvgPicture.asset(
-                                                        DivineIconName
-                                                            .linkSimple
-                                                            .assetPath,
-                                                        width: 24,
-                                                        height: 24,
-                                                        colorFilter:
-                                                            const ColorFilter.mode(
-                                                              VineTheme.primary,
-                                                              BlendMode.srcIn,
-                                                            ),
-                                                      ),
-                                                    ),
-                                                  ),
-                                                ),
-                                              ],
-                                            ),
-                                          ),
-                                        ],
-                                      );
-                                    },
-                                  ),
-                                ),
+                              ProfileAvatarSection(
+                                nameController: _nameController,
                               ),
                               const SizedBox(height: 24),
 
-                              // Display Name
-                              Padding(
-                                padding: const EdgeInsetsDirectional.only(
-                                  start: 16,
-                                ),
-                                child: Text(
-                                  context.l10n.profileSetupDisplayNameLabel,
-                                  style: VineTheme.labelMediumFont(
-                                    color: _nameFocusNode.hasFocus
-                                        ? VineTheme.primary
-                                        : VineTheme.onSurfaceMuted,
-                                  ),
-                                ),
-                              ),
-                              TextFormField(
+                              DisplayNameField(
                                 controller: _nameController,
                                 focusNode: _nameFocusNode,
-                                autovalidateMode:
-                                    AutovalidateMode.onUserInteraction,
-                                style: VineTheme.bodyLargeFont(
-                                  color: VineTheme.onSurface,
-                                ),
-                                decoration: InputDecoration(
-                                  isCollapsed: true,
-                                  hintText:
-                                      context.l10n.profileSetupDisplayNameHint,
-                                  helperText: context
-                                      .l10n
-                                      .profileSetupDisplayNameHelper,
-                                  helperStyle: const TextStyle(
-                                    color: VineTheme.onSurfaceMuted,
-                                    fontSize: 12,
-                                  ),
-                                  hintStyle: _kHintStyle,
-                                  border: _kInputBorder,
-                                  enabledBorder: _kInputBorder,
-                                  focusedBorder: _kInputBorder,
-                                  errorBorder: _kInputBorder,
-                                  focusedErrorBorder: _kInputBorder,
-                                  contentPadding: const EdgeInsets.all(16),
-                                ),
-                                textInputAction: TextInputAction.next,
-                                onFieldSubmitted: (_) =>
-                                    FocusScope.of(context).nextFocus(),
-                                validator: (value) {
-                                  if (value == null || value.trim().isEmpty) {
-                                    return context
-                                        .l10n
-                                        .profileSetupDisplayNameRequired;
-                                  }
-                                  return null;
-                                },
                               ),
                               const SizedBox(height: 16),
 
-                              // Bio
-                              Padding(
-                                padding: const EdgeInsets.symmetric(
-                                  horizontal: 16,
-                                ),
-                                child: Row(
-                                  mainAxisAlignment:
-                                      MainAxisAlignment.spaceBetween,
-                                  children: [
-                                    Text(
-                                      context.l10n.profileSetupBioLabel,
-                                      style: VineTheme.labelMediumFont(
-                                        color: _bioFocusNode.hasFocus
-                                            ? VineTheme.primary
-                                            : VineTheme.onSurfaceMuted,
-                                      ),
-                                    ),
-                                    Text(
-                                      '${_bioController.text.length}/360',
-                                      style: VineTheme.labelMediumFont(
-                                        color: VineTheme.onSurfaceMuted,
-                                      ),
-                                    ),
-                                  ],
-                                ),
-                              ),
-                              TextFormField(
-                                controller: _bioController,
-                                focusNode: _bioFocusNode,
-                                style: VineTheme.bodyLargeFont(
-                                  color: VineTheme.onSurface,
-                                ),
-                                decoration: InputDecoration(
-                                  isCollapsed: true,
-                                  hintText: context.l10n.profileSetupBioHint,
-                                  hintStyle: _kHintStyle,
-                                  border: _kInputBorder,
-                                  enabledBorder: _kInputBorder,
-                                  focusedBorder: _kInputBorder,
-                                  errorBorder: _kInputBorder,
-                                  focusedErrorBorder: _kInputBorder,
-                                  contentPadding: const EdgeInsets.all(16),
-                                  counterText: '',
-                                ),
-                                maxLines: null,
-                                minLines: 1,
-                                maxLength: 360,
-                                textInputAction: TextInputAction.next,
-                                onFieldSubmitted: (_) =>
-                                    FocusScope.of(context).nextFocus(),
-                                onChanged: (_) => setState(() {}),
-                              ),
+                              BioField(controller: _bioController),
                               const SizedBox(height: 16),
 
-                              // Website
-                              Padding(
-                                padding: const EdgeInsetsDirectional.only(
-                                  start: 16,
-                                ),
-                                child: Text(
-                                  context.l10n.profileSetupWebsiteLabel,
-                                  style: VineTheme.labelMediumFont(
-                                    color: _websiteFocusNode.hasFocus
-                                        ? VineTheme.primary
-                                        : VineTheme.onSurfaceMuted,
-                                  ),
-                                ),
-                              ),
-                              TextFormField(
-                                controller: _websiteController,
-                                focusNode: _websiteFocusNode,
-                                style: VineTheme.bodyLargeFont(
-                                  color: VineTheme.onSurface,
-                                ),
-                                decoration: InputDecoration(
-                                  isCollapsed: true,
-                                  hintText:
-                                      context.l10n.profileSetupWebsiteHint,
-                                  hintStyle: _kHintStyle,
-                                  border: _kInputBorder,
-                                  enabledBorder: _kInputBorder,
-                                  focusedBorder: _kInputBorder,
-                                  errorBorder: _kInputBorder,
-                                  focusedErrorBorder: _kInputBorder,
-                                  contentPadding: const EdgeInsets.all(16),
-                                ),
-                                keyboardType: TextInputType.url,
-                                textInputAction: TextInputAction.next,
-                                onFieldSubmitted: (_) =>
-                                    FocusScope.of(context).nextFocus(),
-                              ),
+                              WebsiteField(controller: _websiteController),
                               const SizedBox(height: 16),
 
                               const VerifiedAccountsSection(),
@@ -817,121 +151,7 @@ class _ProfileSetupScreenViewState extends ConsumerState<ProfileSetupScreenView>
                               const PublicKeyLink(),
                               const SizedBox(height: 16),
 
-                              // NIP-05 Username (optional)
-                              BlocBuilder<
-                                ProfileEditorBloc,
-                                ProfileEditorState
-                              >(
-                                buildWhen: (prev, curr) =>
-                                    prev.nip05Mode != curr.nip05Mode,
-                                builder: (context, editorState) {
-                                  final isExternal =
-                                      editorState.nip05Mode ==
-                                      Nip05Mode.external_;
-                                  return Column(
-                                    crossAxisAlignment:
-                                        CrossAxisAlignment.start,
-                                    children: [
-                                      Padding(
-                                        padding: const EdgeInsets.only(
-                                          left: 16,
-                                        ),
-                                        child: Text(
-                                          context
-                                              .l10n
-                                              .profileSetupUsernameLabel,
-                                          style: VineTheme.labelMediumFont(
-                                            color:
-                                                _usernameFocusNode.hasFocus &&
-                                                    !isExternal
-                                                ? VineTheme.primary
-                                                : VineTheme.onSurfaceMuted,
-                                          ),
-                                        ),
-                                      ),
-                                      TextFormField(
-                                        controller: _nip05Controller,
-                                        focusNode: _usernameFocusNode,
-                                        enabled: !isExternal,
-                                        style: VineTheme.bodyLargeFont(
-                                          color: isExternal
-                                              ? VineTheme.onSurfaceMuted
-                                              : VineTheme.onSurface,
-                                        ),
-                                        autovalidateMode:
-                                            AutovalidateMode.onUserInteraction,
-                                        decoration: InputDecoration(
-                                          isCollapsed: true,
-                                          hintText: context
-                                              .l10n
-                                              .profileSetupUsernameHint,
-                                          helperText: context
-                                              .l10n
-                                              .profileSetupUsernameHelper,
-                                          helperStyle: const TextStyle(
-                                            color: VineTheme.onSurfaceMuted,
-                                            fontSize: 12,
-                                          ),
-                                          hintStyle: const TextStyle(
-                                            color: VineTheme.onSurfaceMuted,
-                                          ),
-                                          border: _kInputBorder,
-                                          enabledBorder: _kInputBorder,
-                                          disabledBorder: _kInputBorder,
-                                          focusedBorder: _kInputBorder,
-                                          errorBorder: _kInputBorder,
-                                          focusedErrorBorder: _kInputBorder,
-                                          contentPadding: const EdgeInsets.all(
-                                            16,
-                                          ),
-                                          prefixText: '@',
-                                          prefixStyle: VineTheme.bodyLargeFont(
-                                            color: VineTheme.onSurfaceMuted,
-                                          ),
-                                          suffixText: '.divine.video',
-                                          suffixStyle: VineTheme.bodyLargeFont(
-                                            color: VineTheme.onSurfaceMuted,
-                                          ),
-                                          errorMaxLines: 2,
-                                        ),
-                                        // Lowercase as the user types and
-                                        // restrict to canonical subdomain
-                                        // characters. The name server stores
-                                        // and resolves usernames as lowercase,
-                                        // so normalizing here avoids a
-                                        // confusing "invalid format" error
-                                        // for a typed capital letter.
-                                        inputFormatters: [
-                                          const LowercaseTextInputFormatter(),
-                                          FilteringTextInputFormatter.allow(
-                                            RegExp('[a-z0-9-]'),
-                                          ),
-                                        ],
-                                        textInputAction: TextInputAction.next,
-                                        onFieldSubmitted: (_) =>
-                                            FocusScope.of(context).nextFocus(),
-                                        onChanged: (value) => context
-                                            .read<ProfileEditorBloc>()
-                                            .add(UsernameChanged(value)),
-                                      ),
-                                      // Username status indicators
-                                      if (!isExternal)
-                                        BlocBuilder<
-                                          ProfileEditorBloc,
-                                          ProfileEditorState
-                                        >(
-                                          builder: (context, state) =>
-                                              UsernameStatusIndicator(
-                                                status: state.usernameStatus,
-                                                error: state.usernameError,
-                                                formatMessage:
-                                                    state.usernameFormatMessage,
-                                              ),
-                                        ),
-                                    ],
-                                  );
-                                },
-                              ),
+                              UsernameField(controller: _nip05Controller),
                               const SizedBox(height: 24),
 
                               // Banner section: image upload + color swatches.
@@ -1043,244 +263,6 @@ class _ProfileSetupScreenViewState extends ConsumerState<ProfileSetupScreenView>
     );
   }
 
-  /// Attempts to reconnect relays and re-dispatches [ProfileSaved].
-  ///
-  /// Called from the retry CTA on the no-relays-connected SnackBar. Triggers
-  /// [NostrClient.retryDisconnectedRelays] before re-submitting so the relay
-  /// pool has a chance to reconnect before the publish is attempted again.
-  Future<void> _retryAfterRelayReconnect(
-    BuildContext context,
-    String pubkey,
-  ) async {
-    await ref.read(nostrServiceProvider).retryDisconnectedRelays();
-    if (!context.mounted) return;
-    context.read<ProfileEditorBloc>().add(
-      ProfileSaved(
-        pubkey: pubkey,
-        displayName: _nameController.text,
-        about: _bioController.text,
-        website: _websiteController.text,
-        username: _nip05Controller.text,
-        // Picture and banner sourced from bloc state (see ProfileSaved
-        // dispatch above) via `effectiveBanner` / `effectivePictureUrl`.
-      ),
-    );
-  }
-
-  ImageProvider<Object>? _buildProfilePictureProvider(
-    ProfileEditorState editorState,
-  ) {
-    // Priority:
-    //   1. Local pick preview (only relevant during upload — the bloc has
-    //      no URL yet).
-    //   2. Staged picture from bloc state (post-upload or manual URL).
-    //   3. Persisted picture from bloc state (current kind 0 value).
-    //   4. Placeholder.
-    if (editorState.pendingAvatarStatus == PendingAvatarStatus.uploading) {
-      if (_selectedImageBytes != null) return MemoryImage(_selectedImageBytes!);
-      if (_selectedImage != null) return FileImage(_selectedImage!);
-    }
-
-    final pending = editorState.pendingPictureUrl;
-    if (pending != null && pending.isNotEmpty) {
-      return NetworkImage(pending);
-    }
-
-    final persisted = editorState.persistedPictureUrl;
-    if (persisted != null && persisted.isNotEmpty) {
-      return NetworkImage(persisted);
-    }
-
-    return null;
-  }
-
-  /// Platform-aware image selection.
-  ///
-  /// Native (mobile + desktop): selects an [XFile] with a real filesystem
-  /// path, wraps it in `dart:io File`, and routes through
-  /// `BlossomUploadService.uploadImage` so the platform-channel EXIF
-  /// stripper runs.
-  ///
-  /// Web: `image_picker` returns an [XFile] whose `.path` is a blob URL
-  /// that `dart:io` cannot resolve, so we read the bytes directly and
-  /// route through `BlossomUploadService.uploadImageBytes`, which strips
-  /// EXIF in pure Dart and uploads from memory.
-  Future<void> _pickImage(ImageSource source) async {
-    try {
-      Log.info(
-        '🖼️ Attempting to pick image from ${source.name} on '
-        '${kIsWeb ? "web" : defaultTargetPlatform.name}',
-        name: 'ProfileSetupScreen',
-        category: LogCategory.ui,
-      );
-
-      final picked = await _pickXFile(source);
-      if (picked == null) {
-        Log.info(
-          '❌ No image selected',
-          name: 'ProfileSetupScreen',
-          category: LogCategory.ui,
-        );
-        return;
-      }
-      Log.info(
-        '✅ Image picked: ${picked.name}',
-        name: 'ProfileSetupScreen',
-        category: LogCategory.ui,
-      );
-
-      final pubkey = ref.read(authServiceProvider).currentPublicKeyHex;
-      if (pubkey == null) {
-        Log.error(
-          'Cannot upload avatar: no public key available',
-          name: 'ProfileSetupScreen',
-          category: LogCategory.ui,
-        );
-        return;
-      }
-
-      if (kIsWeb) {
-        // Resolve the blob synchronously here — once we navigate away from
-        // the picker the URL can be revoked.
-        final bytes = await picked.readAsBytes();
-        if (!mounted) return;
-        setState(() {
-          _selectedImage = null;
-          _selectedImageBytes = bytes;
-          _pictureController.clear();
-        });
-        context.read<ProfileEditorBloc>().add(
-          ProfilePictureUploadRequested(
-            pubkey: pubkey,
-            bytes: bytes,
-            filename: picked.name,
-          ),
-        );
-      } else {
-        if (!mounted) return;
-        final file = File(picked.path);
-        setState(() {
-          _selectedImage = file;
-          _selectedImageBytes = null;
-          _pictureController.clear();
-        });
-        context.read<ProfileEditorBloc>().add(
-          ProfilePictureUploadRequested(pubkey: pubkey, file: file),
-        );
-      }
-    } catch (e) {
-      Log.error(
-        'Error picking image: $e',
-        name: 'ProfileSetupScreen',
-        category: LogCategory.ui,
-      );
-
-      // Show user-friendly error message
-      if (mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Text(
-              source == ImageSource.gallery
-                  ? context.l10n.profileSetupImageSelectionFailed
-                  : context.l10n.profileSetupCameraAccessFailed('$e'),
-            ),
-            backgroundColor: VineTheme.error,
-            duration: const Duration(seconds: 5),
-            action: SnackBarAction(
-              label: context.l10n.profileSetupGotItButton,
-              textColor: VineTheme.whiteText,
-              onPressed: () {},
-            ),
-          ),
-        );
-      }
-    }
-  }
-
-  /// Picks a single image, returning the picker's [XFile] without resolving
-  /// it to a `dart:io File`. The returned XFile may have a blob-URL `path`
-  /// on web; callers must use [XFile.readAsBytes] there rather than
-  /// constructing a `File`.
-  Future<XFile?> _pickXFile(ImageSource source) async {
-    // image_picker handles both gallery and camera on web, mobile, and
-    // (since plugin updates) desktop camera. file_selector is only
-    // preferred for native desktop gallery, where it provides a richer
-    // file-type filter UX.
-    if (!kIsWeb && source == ImageSource.gallery && _isDesktopPlatform()) {
-      return _pickXFileFromDesktop();
-    }
-
-    try {
-      return await _picker.pickImage(
-        source: source,
-        maxWidth: 1024,
-        maxHeight: 1024,
-        imageQuality: 85,
-        requestFullMetadata: false,
-      );
-    } catch (e) {
-      Log.error(
-        'image_picker error: $e',
-        name: 'ProfileSetupScreen',
-        category: LogCategory.ui,
-      );
-      rethrow;
-    }
-  }
-
-  /// Check if running on desktop platform.
-  ///
-  /// Always returns false on web — `defaultTargetPlatform` reports the
-  /// host OS in a desktop browser, but a browser is not desktop for
-  /// picker-routing purposes (no real filesystem access).
-  bool _isDesktopPlatform() {
-    if (kIsWeb) return false;
-    return defaultTargetPlatform == TargetPlatform.macOS ||
-        defaultTargetPlatform == TargetPlatform.windows ||
-        defaultTargetPlatform == TargetPlatform.linux;
-  }
-
-  /// Use file_selector for native desktop platforms.
-  Future<XFile?> _pickXFileFromDesktop() async {
-    try {
-      Log.info(
-        '🖥️ Starting desktop file picker...',
-        name: 'ProfileSetupScreen',
-        category: LogCategory.ui,
-      );
-
-      final typeGroup = XTypeGroup(
-        label: context.l10n.profileSetupImagesTypeGroup,
-        extensions: const <String>['jpg', 'jpeg', 'png', 'gif', 'webp', 'bmp'],
-      );
-
-      final file = await openFile(acceptedTypeGroups: <XTypeGroup>[typeGroup]);
-
-      if (file == null) {
-        Log.info(
-          '❌ Desktop file picker: User cancelled or no file selected',
-          name: 'ProfileSetupScreen',
-          category: LogCategory.ui,
-        );
-        return null;
-      }
-
-      Log.info(
-        '✅ Desktop file selected: ${file.name}',
-        name: 'ProfileSetupScreen',
-        category: LogCategory.ui,
-      );
-      return file;
-    } catch (e) {
-      Log.error(
-        'Desktop file picker error: $e',
-        name: 'ProfileSetupScreen',
-        category: LogCategory.ui,
-      );
-      rethrow;
-    }
-  }
-
   void _showNostrInfoSheet(BuildContext context) {
     // Unfocus any field before opening sheet
     FocusScope.of(context).unfocus();
@@ -1294,111 +276,4 @@ class _ProfileSetupScreenViewState extends ConsumerState<ProfileSetupScreenView>
       }
     });
   }
-
-  void _showImageUrlSheet(BuildContext context) {
-    // Unfocus any field before opening sheet
-    FocusScope.of(context).unfocus();
-    VineBottomSheet.show<void>(
-      context: context,
-      scrollable: false,
-      expanded: false,
-      isScrollControlled: true,
-      title: Text(
-        context.l10n.profileSetupImageUrlTitle,
-        style: VineTheme.titleMediumFont(color: VineTheme.onSurface),
-      ),
-      children: [
-        Builder(
-          builder: (sheetContext) => Padding(
-            padding: EdgeInsets.only(
-              left: 16,
-              right: 16,
-              top: 16,
-              bottom: MediaQuery.of(sheetContext).viewInsets.bottom + 16,
-            ),
-            child: TextFormField(
-              controller: _pictureController,
-              style: const TextStyle(color: VineTheme.whiteText),
-              cursorColor: VineTheme.primary,
-              decoration: InputDecoration(
-                hintText: 'https://example.com/image.jpg',
-                hintStyle: const TextStyle(color: VineTheme.lightText),
-                filled: true,
-                fillColor: VineTheme.surfaceContainer,
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(22),
-                  borderSide: BorderSide.none,
-                ),
-              ),
-              textInputAction: TextInputAction.done,
-              onChanged: (_) => setState(() {}),
-              onFieldSubmitted: (_) => Navigator.of(sheetContext).pop(),
-              keyboardType: TextInputType.url,
-              autofocus: true,
-            ),
-          ),
-        ),
-      ],
-    ).then((_) {
-      // Stage the URL the user typed so the avatar widget previews it and
-      // Save can publish it. Empty string clears any prior staged change.
-      if (context.mounted) {
-        context.read<ProfileEditorBloc>().add(
-          ProfilePictureUrlSet(_pictureController.text),
-        );
-        FocusScope.of(context).unfocus();
-      }
-    });
-  }
-}
-
-typedef VerifierRoutePusher =
-    Future<void> Function(String location, {Object? extra});
-
-/// Opens the Divine verifier.
-///
-/// Native in-app WebViews cannot complete the verifier's login flow because it
-/// leaves `verifier.divine.video` for `login.divine.video` and OAuth providers.
-/// Those hand-offs need real browser tabs, cookies, and redirects, so native
-/// platforms launch the system browser. Flutter web keeps the existing iframe
-/// signer bridge because it runs in the browser and does not use
-/// `webview_flutter`. Web refreshes after the iframe route returns; native
-/// refreshes from the screen's app-resume callback.
-@visibleForTesting
-Future<bool> launchVerifierFlow({
-  required ProfileEditorBloc editorBloc,
-  required MyProfileBloc myProfileBloc,
-  VerifierRoutePusher? pushVerifierRoute,
-  bool isWeb = kIsWeb,
-}) async {
-  final verifier = preloadedNostrApps.firstWhere(
-    (app) => app.slug == 'verifier',
-  );
-
-  var launched = false;
-  try {
-    if (isWeb && pushVerifierRoute != null) {
-      await pushVerifierRoute(
-        WebIframeSandboxScreen.pathForAppId(verifier.id),
-        extra: verifier,
-      );
-      launched = true;
-    } else {
-      launched = await launchUrl(
-        Uri.parse(verifier.launchUrl),
-        mode: LaunchMode.externalApplication,
-      );
-    }
-  } catch (error) {
-    UnifiedLogger.warning(
-      'Failed to open Divine verifier: $error',
-      name: 'ProfileSetupScreen',
-    );
-  }
-
-  editorBloc.add(const VerifierLaunchHandled());
-  if (launched && isWeb) {
-    myProfileBloc.add(const MyProfileFetchRequested());
-  }
-  return launched;
 }
