+++ mobile/lib/screens/profile_setup/widgets/profile_setup_field_decorations.dart
@@ -1,0 +1,11 @@
+import 'package:divine_ui/divine_ui.dart';
+import 'package:flutter/material.dart';
+
+/// Shared underline border for the profile-setup form fields.
+const profileFieldBorder = UnderlineInputBorder(
+  borderRadius: BorderRadius.zero,
+  borderSide: BorderSide(color: VineTheme.neutral10),
+);
+
+/// Shared hint text style for the profile-setup form fields.
+const profileFieldHintStyle = TextStyle(color: VineTheme.lightText);
