+++ mobile/lib/screens/profile_setup/widgets/widgets.dart
@@ -1,5 +1,14 @@
 export 'banner_editing_block.dart';
+export 'bio_field.dart';
+export 'display_name_field.dart';
+export 'profile_avatar_section.dart';
+export 'profile_image_picker.dart';
+export 'profile_setup_field_decorations.dart';
+export 'profile_setup_listeners.dart';
 export 'profile_setup_upload_errors.dart';
 export 'public_key_link.dart';
 export 'save_button.dart';
+export 'username_field.dart';
 export 'verified_accounts_section.dart';
+export 'verifier_flow.dart';
+export 'website_field.dart';
